The canonicalization of request URLs alphabetizes the parameters
The parameters do not retain their original ordering in the URLs. The URLCanonicalizer references wikipedia, which states
that the lexicographical ordering of parameters can cause problems (it has for me), yet the canonicalizer still
lexicographically orders the parameters.

I have changed all instances of SortedMap during the url parameter conocalization
process with a LinkedHashMap, which retains the ordering of Entrys inserted.

I can not think of any disadvantages of retaining the original ordering of the url parameters.